Split problem tests into own file in pygmo.
